+++ docs/changelog.rst
@@ -17,6 +17,7 @@ Internal Changes
 ================
 - Skip ``twine check`` test on macOS until shell kill error is fixed. By `Kyle Brindley`_.
 - Restructure internal deployment around Gitlab-CI tag pipelines. By `Kyle Brindley`_.
+- Preserve QOI datatree structure when merging (:merge:`1205`). By `Matthew Fister`_.
 
 *******************
 0.13.4 (2025-05-29)